Ödev yüklendi

+++ usHesabiOdevi_w4.java
@@ -5,15 +5,15 @@ public class usHesabiOdevi_w4 {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
 
-        int a,b;
+        long a,b;
 
         System.out.print("Ussu alinacak sayiyi giriniz : ");
         a = input.nextInt();
         System.out.print("Kac kere us alacaginizi giriniz : ");
         b=input.nextInt();
 
-        int total=1;
-        for (int i =1; i<=b; i++){
+        long total=1;
+        for (long i =1; i<=b; i++){
             total *= a;
         }
